requirements.txt
numpy
trimesh
tqdm
Pillow
scipy
plyfile
pandas
blosc2